Moved event-sourcing code to separate assembly.
