+++ src/Orleans.EventSourcing/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
 Project Orleans Cloud Service SDK ver. 1.0
  
 Copyright (c) Microsoft Corporation
@@ -21,7 +21,6 @@ OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHE
 TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-using System;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -29,8 +28,8 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Orleans")]
-[assembly: AssemblyDescription("Orleans - Orleans API")]
+[assembly: AssemblyTitle("Orleans EventSourcing")]
+[assembly: AssemblyDescription("Orleans - Event Sourcing API")]
 [assembly: AssemblyConfiguration("")]
 
 // Setting ComVisible to false makes the types in this assembly not visible 
@@ -39,7 +38,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("c849bbea-5e7c-456f-9367-664ce98a3ab6")]
+[assembly: Guid("cf7ee78b-39e1-4660-beaf-0e2606098f8c")]
 
 [assembly: InternalsVisibleTo("ClientGenerator, PublicKey=0024000004800000940000000602000000240000525341310004000001000100eb81539dc2728946d4860bad0fbc91ea23d7a835e00547aa4e9a7abf185535b564d29dbe77915b779951fef2f4b1ab39cc3a246cf572c7b217b5c892835a687eb22638183d66930039ff9fa6e7e71cb880f854fb76b884dfc905bf1f3c623abad7d4732c1ce62efff937155e352170831239acbe103a41ed363a90001d3108ca")]
 [assembly: InternalsVisibleTo("OrleansCounterControl, PublicKey=0024000004800000940000000602000000240000525341310004000001000100eb81539dc2728946d4860bad0fbc91ea23d7a835e00547aa4e9a7abf185535b564d29dbe77915b779951fef2f4b1ab39cc3a246cf572c7b217b5c892835a687eb22638183d66930039ff9fa6e7e71cb880f854fb76b884dfc905bf1f3c623abad7d4732c1ce62efff937155e352170831239acbe103a41ed363a90001d3108ca")]
@@ -52,4 +51,3 @@ using System.Runtime.InteropServices;
 [assembly: InternalsVisibleTo("UnitTestGrainInterfaces, PublicKey=0024000004800000940000000602000000240000525341310004000001000100eb81539dc2728946d4860bad0fbc91ea23d7a835e00547aa4e9a7abf185535b564d29dbe77915b779951fef2f4b1ab39cc3a246cf572c7b217b5c892835a687eb22638183d66930039ff9fa6e7e71cb880f854fb76b884dfc905bf1f3c623abad7d4732c1ce62efff937155e352170831239acbe103a41ed363a90001d3108ca")]
 [assembly: InternalsVisibleTo("UnitTestGrains, PublicKey=0024000004800000940000000602000000240000525341310004000001000100eb81539dc2728946d4860bad0fbc91ea23d7a835e00547aa4e9a7abf185535b564d29dbe77915b779951fef2f4b1ab39cc3a246cf572c7b217b5c892835a687eb22638183d66930039ff9fa6e7e71cb880f854fb76b884dfc905bf1f3c623abad7d4732c1ce62efff937155e352170831239acbe103a41ed363a90001d3108ca")]
 [assembly: InternalsVisibleTo("TesterInternal, PublicKey=0024000004800000940000000602000000240000525341310004000001000100eb81539dc2728946d4860bad0fbc91ea23d7a835e00547aa4e9a7abf185535b564d29dbe77915b779951fef2f4b1ab39cc3a246cf572c7b217b5c892835a687eb22638183d66930039ff9fa6e7e71cb880f854fb76b884dfc905bf1f3c623abad7d4732c1ce62efff937155e352170831239acbe103a41ed363a90001d3108ca")]
-
